+++ jumi-actors-maven-plugin/src/it/compiles-using-project-language-level/src/main/java/example/JavaSevenUser.java
@@ -1,0 +1,12 @@
+// Copyright © 2011-2012, Esko Luontola <www.orfjackal.net>
+// This software is released under the Apache License 2.0.
+// The license text is at http://www.apache.org/licenses/LICENSE-2.0
+
+package example;
+
+import java.util.*;
+
+public class JavaSevenUser {
+
+    public List<String> usesDiamondOperator = new ArrayList<>();
+}